+++ Sources/ConfigDirector/Internal/Constants.swift
@@ -2,6 +2,6 @@ import Foundation
 
 enum Constants {
     static let sdkName = "swift-client-sdk"
-    static let sdkVersion = "1.3.0"
+    static let sdkVersion = "1.4.0"
     static let clientBaseURL = URL(string: "https://client-sdk-api.configdirector.com")!
 }